+++ src/application/context/dependencies/search/SearchDependenciesResponse.ts
@@ -1,0 +1,9 @@
+/**
+ * SearchDependenciesResponse - Response DTO for the search-dependencies use case.
+ */
+
+import { DependencyView } from "../DependencyView.js";
+
+export interface SearchDependenciesResponse {
+  readonly dependencies: DependencyView[];
+}