+++ tests/application/context/dependencies/list/ListDependenciesQueryHandler.test.ts
@@ -15,6 +15,7 @@ describe("GetDependenciesQueryHandler", () => {
     mockReader = {
       findAll: jest.fn(),
       findByIds: jest.fn(),
+      search: jest.fn(),
     } as jest.Mocked<IDependencyViewReader>;
 
     queryHandler = new GetDependenciesQueryHandler(mockReader);